requirements.txt
fastapi
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
sqlmodel
uvicorn

#for testing
pytest
pytest-cov
requests